backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic==0.18.1",
    "cryptography==42.0.0",
    "fastapi==0.109.0",
    "httpx==0.26.0",
    "openai==1.12.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "pymysql==1.1.0",
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.25",
    "tenacity==8.2.3",
    "uvicorn[standard]==0.27.0",
]
